Extract s3reader
